Refactor: MainActivity.kt -> HouseKeeperActivity.kt로 이름 변경 & ui 폴더 내로 이동 & PrefsManager.kt util 폴더 내로 이동
MainFragment.kt와 연결된 MainViewModel과 MainActivity.kt와 연결된 MainViewModel이 존재하여 중복된 이름 피하고자 이름 변경
