+++ app/src/main/java/com/depromeet/housekeeper/service/FCMWorker.kt
@@ -4,7 +4,7 @@ import android.content.Context
 import androidx.work.CoroutineWorker
 import androidx.work.ForegroundInfo
 import androidx.work.WorkerParameters
-import com.depromeet.housekeeper.local.PrefsManager
+import com.depromeet.housekeeper.util.PrefsManager
 import com.depromeet.housekeeper.model.Token
 import com.depromeet.housekeeper.network.remote.repository.Repository
 import kotlinx.coroutines.flow.collect
