+++ app/src/main/java/com/depromeet/housekeeper/ui/HouseKeeperViewModel.kt
@@ -1,4 +1,4 @@
-package com.depromeet.housekeeper
+package com.depromeet.housekeeper.ui
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
@@ -8,7 +8,7 @@ import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.launch
 
 
-class MainViewModel : ViewModel(){
+class HouseKeeperViewModel : ViewModel(){
     private val _isLoading: MutableStateFlow<Boolean> = MutableStateFlow(true)
     val isLoading : StateFlow<Boolean>
     get() = _isLoading
