+++ app/src/main/java/com/depromeet/housekeeper/ui/join/GroupInfoViewModel.kt
@@ -2,7 +2,7 @@ package com.depromeet.housekeeper.ui.join
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.depromeet.housekeeper.local.PrefsManager
+import com.depromeet.housekeeper.util.PrefsManager
 import com.depromeet.housekeeper.model.Assignee
 import com.depromeet.housekeeper.network.remote.repository.Repository
 import kotlinx.coroutines.flow.MutableStateFlow
