+++ app/src/main/java/com/depromeet/housekeeper/ui/signIn/LoginFragment.kt
@@ -13,7 +13,7 @@ import androidx.navigation.fragment.findNavController
 import androidx.navigation.fragment.navArgs
 import com.depromeet.housekeeper.R
 import com.depromeet.housekeeper.databinding.FragmentLoginBinding
-import com.depromeet.housekeeper.local.PrefsManager
+import com.depromeet.housekeeper.util.PrefsManager
 import com.depromeet.housekeeper.model.enums.SignViewType
 import com.depromeet.housekeeper.network.remote.model.LoginResponse
 import com.google.android.gms.auth.api.signin.GoogleSignIn
