+++ app/src/main/res/layout/activity_main.xml
@@ -6,7 +6,7 @@
   <androidx.constraintlayout.widget.ConstraintLayout
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    tools:context=".ui.HouseKeeperActivity">
     
     <androidx.fragment.app.FragmentContainerView
       android:id="@+id/nav_host_fragment"
